custom-hooks done. moved getting media to separate hook
