+++ src/views/Home.jsx
@@ -1,37 +1,13 @@
-import {useEffect, useState} from 'react';
+import {useState} from 'react';
 import MediaRow from '../components/MediaRow';
 import SingleView from '../components/SingleView';
-import {fetchData} from '../utils/fetchData.js';
+
+import useMedia from '../hooks/apiHooks.js';
 
 const Home = () => {
-  const [mediaArray, setMediaArray] = useState([]);
-
-  const getMedia = async () => {
-    try {
-      const mediaData = await fetchData(
-        import.meta.env.VITE_MEDIA_API + '/media',
-      );
-      const url = import.meta.env.VITE_AUTH_API + '/users/';
-      console.log(url);
-      const newData = await Promise.all(
-        mediaData.map(async (item) => {
-          const result = await fetchData(url + item.user_id);
-          return {...item, username: result.username};
-        }),
-      );
-      setMediaArray(newData);
-    } catch (error) {
-      console.error('error', error);
-    }
-  };
-
-  useEffect(() => {
-    getMedia();
-  }, []);
-  console.log(mediaArray);
-
   const [selectedItem, setSelectedItem] = useState(null);
-  console.log('selectedItem', selectedItem);
+  const mediaArray = useMedia();
+
   return (
     <>
       <h2>My Media</h2>
